+++ Algebras/CRel/Operators/CConsume.cpp
@@ -27,7 +27,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "AttrArray.h"
 #include "CRel.h"
 #include "CRelTC.h"
-#include <cstddef>
+#include <cstdint>
 #include <exception>
 #include "ListUtils.h"
 #include "LogMsg.h"
@@ -84,7 +84,7 @@ const OperatorInfo CConsume::info = OperatorInfo(
 ListExpr CConsume::TypeMapping(ListExpr args)
 {
   //Two args?
-  const size_t argCount = nl->ListLength(args);
+  const uint64_t argCount = nl->ListLength(args);
 
   if (argCount < 2 || argCount > 3)
   {
